Backport deployment task support to Bamboo 9.x
Creates a Bamboo 9.6.5-compatible line of the plugin that includes the
deployment-project task type added in b75c92b (RTECO-858), the admin URL
namespace fix from 7dc86e3, and the security improvements from bff5c83 —
all restored to work with the pre-10.x runtime.

Reverts 0ae9650 (Bamboo 10.x compatibility):
  * pom: bamboo 9.6.5, Java 11, amps 8.13.6, spring-scanner 3.0.3
  * pom: drop OSGi <instructions>, restore <enableQuickReload>/
    <extractDependencies>, un-provide the Jackson deps so they bundle
  * ServerConfigManager: back to a concrete @Component class exposing
    getInstance() via ContainerManager.autowireComponent; folds in the
    bff5c83 log-sanitization fixes; ServerConfigManagerImpl deleted
  * spring-scanner.xml: deleted (not needed by scanner 3.x)
  * atlassian-plugin.xml: restore <component> declaration for
    ServerConfigManager (kept /admin/jfrog namespace + JfDeploymentTask)
  * Actions and tasks: drop @Inject/@ComponentImport annotations;
    JfTask/JfDeploymentTask fetch getInstance() lazily so unit tests can
    still inject mocks via the existing setters

Version bumped to 1.0.5.1 so Bamboo 9.x installations pick this over
the 10.x-only 1.0.6+ releases.
