+++ pom.xml
@@ -6,7 +6,7 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>org.jfrog.bamboo</groupId>
     <artifactId>bamboo-jfrog-plugin</artifactId>
-    <version>1.0.8</version>
+    <version>1.0.5.1</version>
 
     <organization>
         <name>JFrog Ltd</name>
@@ -30,14 +30,14 @@
     <packaging>atlassian-plugin</packaging>
 
     <properties>
-        <bamboo.version>10.2.0</bamboo.version>
+        <bamboo.version>9.6.5</bamboo.version>
         <bamboo.data.version>${bamboo.version}</bamboo.data.version>
-        <amps.version>9.1.4</amps.version>
-        <atlassian.spring.scanner.version>5.0.2</atlassian.spring.scanner.version>
+        <amps.version>8.13.6</amps.version>
+        <atlassian.spring.scanner.version>3.0.3</atlassian.spring.scanner.version>
         <!-- This property ensures consistency between the key in atlassian-plugin.xml and the OSGi bundle's key. -->
         <atlassian.plugin.key>${project.groupId}.${project.artifactId}</atlassian.plugin.key>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>17</java.version>
+        <java.version>11</java.version>
         <maven.compiler.source>${java.version}</maven.compiler.source>
         <maven.compiler.target>${java.version}</maven.compiler.target>
         <!-- https://mvnrepository.com/artifact/org.jfrog.buildinfo/build-info-extractor -->
@@ -153,19 +153,11 @@
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-core</artifactId>
             <version>2.17.2</version>
-            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
             <version>2.17.2</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-annotations</artifactId>
-            <version>2.17.2</version>
-            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>org.mockito</groupId>
@@ -191,19 +183,8 @@
                 <configuration>
                     <productVersion>${bamboo.version}</productVersion>
                     <productDataVersion>${bamboo.data.version}</productDataVersion>
-                    <instructions>
-                        <Atlassian-Plugin-Key>${atlassian.plugin.key}</Atlassian-Plugin-Key>
-                        <Bundle-SymbolicName>${atlassian.plugin.key}</Bundle-SymbolicName>
-                        <Spring-Context>*</Spring-Context>
-                        <Export-Package>
-                        </Export-Package>
-                        <Import-Packages>
-                            !com.opensymphony.*,
-                            !org.objectweb.asm.*,
-                            *;resolution:="optional",
-                        </Import-Packages>
-                        <skipManifestValidation>true</skipManifestValidation>
-                    </instructions>
+                    <enableQuickReload>true</enableQuickReload>
+                    <extractDependencies>false</extractDependencies>
                 </configuration>
             </plugin>
             <plugin>
