+++ src/main/java/org/jfrog/bamboo/JfContext.java
@@ -11,7 +11,6 @@ import org.jfrog.bamboo.config.ServerConfig;
 import org.jfrog.bamboo.config.ServerConfigManager;
 import org.jfrog.bamboo.utils.ExecutableRunner;
 
-import javax.inject.Inject;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -28,14 +27,11 @@ public class JfContext extends AbstractTaskConfigurator {
     public static final String JF_TASK_COMMAND = "jf.task.command";
     public static final String JF_TASK_WORKING_DIRECTORY = "jf.task.working.directory";
 
-    @Inject
-    private ServerConfigManager serverConfigManager;
-
     @Override
     public void populateContextForCreate(@NotNull Map<String, Object> context) {
         super.populateContextForCreate(context);
         context.put(JF_TASK_COMMAND, "jf ");
-        context.put("serverConfigManager", serverConfigManager);
+        context.put("serverConfigManager", ServerConfigManager.getInstance());
         context.put("selectedServerId", 1);
     }
 
@@ -46,7 +42,7 @@ public class JfContext extends AbstractTaskConfigurator {
         context.put(JF_TASK_SERVER_ID, config.get(JF_TASK_SERVER_ID));
         context.put(JF_TASK_COMMAND, config.get(JF_TASK_COMMAND));
         context.put(JF_TASK_WORKING_DIRECTORY, config.get(JF_TASK_WORKING_DIRECTORY));
-        context.put("serverConfigManager", serverConfigManager);
+        context.put("serverConfigManager", ServerConfigManager.getInstance());
     }
 
     @Override
