+++ src/main/java/org/jfrog/bamboo/JfTask.java
@@ -10,19 +10,16 @@ import com.atlassian.bamboo.v2.build.BuildContext;
 import com.atlassian.bamboo.variable.CustomVariableContext;
 import com.atlassian.bamboo.variable.VariableDefinitionContext;
 import com.atlassian.plugin.PluginAccessor;
-import com.atlassian.plugin.spring.scanner.annotation.imports.ComponentImport;
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.lang.exception.ExceptionUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jfrog.bamboo.config.ServerConfigManager;
 import org.jfrog.bamboo.config.ServerConfig;
-import org.jfrog.bamboo.config.ServerConfigManagerImpl;
 import org.jfrog.bamboo.utils.BambooUtils;
 import org.jfrog.bamboo.utils.BuildLog;
 import org.jfrog.bamboo.utils.ExecutableRunner;
 import org.jfrog.bamboo.utils.Utils;
 
-import javax.inject.Inject;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -38,18 +35,10 @@ import java.util.stream.Collectors;
  */
 public class JfTask extends JfContext implements TaskType {
     private BuildLog buildLog;
-    @Inject
     private ServerConfigManager serverConfigManager;
     private ExecutableRunner commandRunner;
-    @Inject
-    @ComponentImport
     private CustomVariableContext customVariableContext;
-    @Inject
-    @ComponentImport
     private PluginAccessor pluginAccessor;
-
-    @Inject
-    @ComponentImport
     private AdministrationConfigurationAccessor administrationConfigurationAccessor;
 
     /**
@@ -61,6 +50,9 @@ public class JfTask extends JfContext implements TaskType {
     @Override
     public @NotNull TaskResult execute(final @NotNull TaskContext taskContext) {
         buildLog = new BuildLog(taskContext.getBuildLogger());
+        if (serverConfigManager == null) {
+            serverConfigManager = ServerConfigManager.getInstance();
+        }
         ConfigurationMap confMap = taskContext.getConfigurationMap();
         TaskResultBuilder resultBuilder = TaskResultBuilder.newBuilder(taskContext);
         ServerConfig selectedServerConfig = serverConfigManager.getServerConfigById(confMap.get(JF_TASK_SERVER_ID));
