--- src/main/resources/META-INF/spring/spring-scanner.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:atlassian-scanner="http://www.atlassian.com/schema/atlassian-scanner/2"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans
-           http://www.springframework.org/schema/beans/spring-beans-2.5.xsd
-           http://www.atlassian.com/schema/atlassian-scanner/2
-           http://www.atlassian.com/schema/atlassian-scanner/2/atlassian-scanner.xsd">
-    <atlassian-scanner:scan-indexes/>
-</beans>
